remove duplicate LinkPreview.swift
